Use `RCE` from `plottr_components`
Use ColorPickerColor from `plottr_components`

Use `ItemsManagerModal` from `plottr_components`

Put all dependencies into dev dependencies

Library doesn't package them.  It expects dependants to deal with
them.

Fix duplicate React problem & add plugin to detect duplicate modules

Use `PlottrModal` from `plottr_components`

Use EditAttribute from `plottr_components`

Fix tests which failed after moving the RCE
