+++ lib/plottr_components/src/components/rce/Element.js
@@ -34,6 +34,8 @@ const Element = ({ attributes, children, element, openExternal }) => {
           {...attributes}
           title={element.url}
           href={element.url}
+          rel="noreferrer"
+          target="_blank"
           onClick={() => openExternal(element.url)}
         >
           {children}
