+++ lib/plottr_components/src/components/rce/RichTextEditor.js
@@ -46,6 +46,7 @@ const RichTextEditor = ({ openExternal, ...props }) => {
   const [editorWrapperRef, setEditorWrapperRef] = useState(null)
   const key = useRef(Math.random().toString(16))
   const toolbarRef = useRef(null)
+  const bottomTextAreaRef = useRef(null)
   const [showColorPicker, toggleColorPicker] = useState(false)
   const changeColor = (color) => {
     addColorMark(editor, color)
@@ -68,6 +69,39 @@ const RichTextEditor = ({ openExternal, ...props }) => {
     if (value !== newVal) {
       props.onChange(newVal)
     }
+  }
+
+  const scrollToEnd = () => {
+    if (bottomTextAreaRef.current) {
+      bottomTextAreaRef.current.scrollIntoView({
+        block: 'end',
+      })
+    }
+  }
+
+  const handleKeyDown = (event) => {
+    for (const hotkey in HOTKEYS) {
+      if (isHotkey(hotkey, event)) {
+        event.preventDefault()
+        const mark = HOTKEYS[hotkey]
+        toggleMark(editor, mark)
+      }
+    }
+    scrollToEnd()
+  }
+
+  const handleClickEditable = (event) => {
+    if (!editorWrapperRef) return
+    if (editorWrapperRef.firstChild.contains(event.target)) return
+
+    // Focus the Editable content
+    // Select all text
+    // Push cursor/focus to last char
+    // Scroll the view to bottom area of Editable
+    editorWrapperRef.firstChild.focus()
+    document.execCommand('selectAll', false, null)
+    document.getSelection().collapseToEnd()
+    scrollToEnd()
   }
 
   const otherProps = {
@@ -102,6 +136,7 @@ const RichTextEditor = ({ openExternal, ...props }) => {
             </ButtonGroup>
             <Overlay show={showColorPicker} placement="bottom" container={() => toolbarRef.current}>
               <MiniColorPicker
+                darkMode={props.darkMode}
                 chooseColor={changeColor}
                 el={toolbarRef}
                 close={() => toggleColorPicker(false)}
@@ -115,11 +150,7 @@ const RichTextEditor = ({ openExternal, ...props }) => {
             registerEditor(e && e.firstChild)
             setEditorWrapperRef(e)
           }}
-          onClick={(event) => {
-            if (!editorWrapperRef) return
-            if (editorWrapperRef.firstChild.contains(event.target)) return
-            editorWrapperRef.firstChild.focus()
-          }}
+          onClick={(event) => handleClickEditable(event)}
           className={cx('slate-editor__editor', { darkmode: props.darkMode })}
         >
           <Editable
@@ -128,16 +159,9 @@ const RichTextEditor = ({ openExternal, ...props }) => {
             renderLeaf={renderLeaf}
             renderElement={renderElement}
             placeholder={i18n('Enter some text...')}
-            onKeyDown={(event) => {
-              for (const hotkey in HOTKEYS) {
-                if (isHotkey(hotkey, event)) {
-                  event.preventDefault()
-                  const mark = HOTKEYS[hotkey]
-                  toggleMark(editor, mark)
-                }
-              }
-            }}
+            onKeyDown={(event) => handleKeyDown(event)}
           />
+          <div ref={bottomTextAreaRef} />
         </div>
       </div>
     </Slate>
