+++ src/app/components/places/placeView.js
@@ -12,12 +12,11 @@ import {
   Glyphicon,
 } from 'react-bootstrap'
 import { t as i18n } from 'plottr_locales'
-import RichText from '../rce/RichText'
 import ImagePicker from '../images/ImagePicker'
 import Image from '../images/Image'
 import SelectList from '../selectList'
 import BookSelectList from '../project/BookSelectList'
-import { DeleteConfirmModal } from 'connected-components'
+import { DeleteConfirmModal, RichText } from 'connected-components'
 import { actions, selectors } from 'pltr/v2'
 
 const { sortedTagsSelector } = selectors
