+++ src/app/components/timeline/Card.js
@@ -5,7 +5,7 @@ import { bindActionCreators } from 'redux'
 import CardDialog from 'components/timeline/CardDialog'
 import { Popover, OverlayTrigger } from 'react-bootstrap'
 import TagLabel from 'components/tagLabel'
-import RichText from '../rce/RichText'
+import { RichText } from 'connected-components'
 import cx from 'classnames'
 import { FaCircle } from 'react-icons/fa'
 import { helpers, actions, selectors } from 'pltr/v2'
