+++ src/app/components/timeline/CardDialog.js
@@ -17,13 +17,13 @@ import {
 import { t as i18n } from 'plottr_locales'
 import cx from 'classnames'
 import tinycolor from 'tinycolor2'
-import RichText from '../rce/RichText'
 import SelectList from 'components/selectList'
 import {
-  PlottrModal,
   ColorPickerColor,
   DeleteConfirmModal,
   EditAttribute,
+  PlottrModal,
+  RichText,
 } from 'connected-components'
 import { helpers, actions, selectors } from 'pltr/v2'
 import MiniColorPicker from '../MiniColorPicker'
