+++ src/app/index.js
@@ -19,7 +19,7 @@ import log from 'electron-log'
 import Modal from 'react-modal'
 import SETTINGS from '../common/utils/settings'
 import { ActionCreators } from 'redux-undo'
-import editorRegistry from './components/rce/editor-registry'
+import { editorRegistry } from 'connected-components'
 import { setupI18n } from 'plottr_locales'
 import { displayFileName, editKnownFilePath } from '../common/utils/known_files'
 import { addNewCustomTemplate } from '../common/utils/custom_templates'
